refactor(check): split output reporting modules
